SDK, reorganize mod structure for clarity (#4749)
* SDK, reorganize mod structure for clarity

Moves a couple of utility functions from the root level into a pub channel and utils mod. Should make the docs a lot easier to read. Also aligns the crate file structure with its pub interface for clarity.

This is technically a breaking change, but all moved functions are very new, and I don't expect teams to already be depending on them. And if they are, the compiler error message should make migration not too difficult.

* fix warnings

* don't squash old core modules together

* fix typo

* improve doc comment

* pub use the log crate

* fix typo

* move attestation structs to submodules

* fix typo

* ensure mock structs are documented

* appease clippy

* appease clippy 2