Allow batching lookup data from Wasm (#4761)
Adds a new microRPC method for multiple lookups. In the future the two may be easily unified, but at the moment I didn't want to risk breaking existing functionality.

Add a new Wasm module to perform multiple lookups. The difference from the existing one is that it deals with protos as input / output instead of just the raw value to look up, and also a parameter to determine how to do the lookups (to help with benchmarking).

Remove randomness in benchmark (it was not actually doing anything useful in the way it was written)