Switch to custom hash table to save memory.
The hashbrown table, when using (Bytes, Bytes) as backing, uses far too much memory when keys and values are small.  This custome hash table reduces overhead to about 21 bytes per key/value pair.  It is optimized for large tables that bust out of on-chip cache.  A swiss table is used to reduce cache misses on average to 2, one for indexing into the table, and one to index into the vector backing the keys/value pairs.