Switch to custom hash table to save memory. (#4805)
* Switch to custom hash table to save memory.

The hashbrown table, when using (Bytes, Bytes) as backing, uses far too much memory when keys and values are small.  This custome hash table reduces overhead to about 21 bytes per key/value pair.  It is optimized for large tables that bust out of on-chip cache.  A swiss table is used to reduce cache misses on average to 2, one for indexing into the table, and one to index into the vector backing the keys/value pairs.

* Switch to custom hash table to save memory.

The hashbrown table, when using (Bytes, Bytes) as backing, uses far too much memory when keys and values are small.  This custome hash table reduces overhead to about 21 bytes per key/value pair.  It is optimized for large tables that bust out of on-chip cache.  A swiss table is used to reduce cache misses on average to 2, one for indexing into the table, and one to index into the vector backing the keys/value pairs.

* Got most tests working with the new hash table.

* Add ability for hash table to grow, since both tests, and he old weather loookup app do not like to pre-allocate data.

* Reformatted 2 files

* Reverted two kernel configs that were mistakenly committed.

* Chunk hash table data on the fly so we don't have to reserve it.

* Compress lengths so that lenths < 64 are encoded as one byte.

* Compress lengths so that lenths < 64 are encoded as one byte.

* Ran rustfmt on lookup_htbl.rs

* Fixed a write_len bug and added more tests.

* Ran rustfmt again

* More format fixes

* Return an error rather than panic when we detect a duplicate key.

* Reduced files changed, fixed server test.

There is a public datatpe from lookup.rs called Data, which is used externally only to pass data to the lookup module.  This type is used as a Vec<(Bytes, Bytes)> everywhere.  This CL changes Data from HashMap to Vec<(Bytes, Bytes)> for use in external tests.

* Reduced chunk size to 1GiB from 4GiB to save time on unit tests.

* Fixed warnings about unused results.

* Fixes to make Clippy happy.

* Got clippy to pass.

* Reduced chunk size from 1GiB to 2MiB.

* Reduce memory used in unit tests to fit in a VM with only 256MiB

* Simplified map_err in two places.

* Remove Data type from lookup.rs and renamed DataInternal to Data.

The public declaration of hte Data type seemed to cause confusion when writing tests.  In reality, Data was only ever used as a container for (Bytes, Bytes).  Tests were changed to used Vec<(Bytes, Bytes)>, which is more readable.

* Allow duplicate keys to be inserted without returning an error.

This matches more closely the behavior of HashMap, and should make it simpler to switch from LookupHtbl to HashMap and back.

* Allow duplicate keys to be inserted without returning an error.

This matches more closely the behavior of HashMap, and should make it simpler to switch from LookupHtbl to HashMap and back.

* Adding rustdoc comments.

* Changed main hash table to use Entry {hash_byte: u8, data_index: [u8; 5]}.

This makes it a bit more readable, and maybe the compiler might optimize a bit better.

* Added several comments to the code.

* Ran rustfmt on lookup_htbl.rs

* Disable a poor clippy warning.

It want's me to replace a for 0..N loop with a much slower iterator.  Um... no.

* Fixed wasm_benchmark.rs

* Reduced lookup loop complexity and cleaned up a few other places to improve readability.

* Tuned the hash funciton a bit an improved the cycle length calculator

* Replaced 5 with INDEX_SIZE.

* Add a non-reversible operation in Rand to clobber group structure.

* Fixed clippy warnings.