Added randomness to lookup_htbl.
This makes it harder to extract a user's location via cache-timing, but not impossible.  The hash function is extended to 2 rounds and is turned into a hash rather than a pseudo-random permutation.  This modified 2-round hash passes the dieharder tests when feeding it with a counter, and outputting the lwoer 32 bits of every hash.  This is more than enough proof that the hash function output is suitable for a hash table.