Reduced hashing back down to 1 round.
This no longer passes dieharder tests, but all we really want is a nice random-ish distribution that does not have more collisions in the hash table than expected.  We now measure this in a collision test, and one round is good enough.  This is a both faster and a better distribution than the ahash::RandomState algorithm.