Add tool to predict kernel boot measurements (#4835)
When we start a new VM we pass the kernel in a bzImage format, but the VMM splits it into 2 parts before passing it to the Stage 0 firmware. So the measurements by stage 0 are of these two derived parts.

This tool applies the same changes to that we can predict what both measurements would be for a specific bzImage kernel.