Added randomness to lookup_htbl. (#4839)
* Added randomness to lookup_htbl.

This makes it harder to extract a user's location via cache-timing, but not impossible.  The hash function is extended to 2 rounds and is turned into a hash rather than a pseudo-random permutation.  This modified 2-round hash passes the dieharder tests when feeding it with a counter, and outputting the lwoer 32 bits of every hash.  This is more than enough proof that the hash function output is suitable for a hash table.

* Updated Cargo.lock file.

* Updated Cargo.lock file.

* Just reformatted 1 file.

* Reduced hashing back down to 1 round.

This no longer passes dieharder tests, but all we really want is a nice random-ish distribution that does not have more collisions in the hash table than expected.  We now measure this in a collision test, and one round is good enough.  This is a both faster and a better distribution than the ahash::RandomState algorithm.

* Blocked clippy from trying to make an iterator to traverse the hash table.