Slight improvement to the hash fucntion.
By XORing the input value into the outpout only at the end, we maintain a permutation until that step.  This ensures no information is lost in the intermediate steps.  With this scheme, a 3-round version passes the diehardet tests when ouitputing all bits of the hash when the input is a counter.