Enhanced hash_u64 as 2-rounds. (#4852)
* Enhanced hash_u64 as 2-rounds.

* Slight improvement to the hash fucntion.

By XORing the input value into the outpout only at the end, we maintain a permutation until that step.  This ensures no information is lost in the intermediate steps.  With this scheme, a 3-round version passes the diehardet tests when ouitputing all bits of the hash when the input is a counter.

* Switched to folding multiply, similar to aHash, but with some more mixing before.

This now passes distribution tests, cycle length tests, and is just as fast as aHash's fallback path for u64.  The current aHash code still fails distribution tests, but if the author adopts something like what is in this PR, we could switch to aHash.