Fail on non-empty reference values (#4850)
Two goals achieved here:

(1) Make the verification fail on non-empty firmware version. We don't have a spec for that reference value yet. It would be good to address this in one go for all of AttestationReportValues.

(2) Make the verification fail for non-skip stage0 reference value. This adds code that could already work with the derived endorsements but fails now since we don't have the derived endorsements for stage0.

See b/327069120