Change the reduce function to use 128-bit multiply.
The reduce function before this change did not uniformly select a hash table location, slightly increasing the number of collisions when inserting a billion entries into the table.  As the number of keys approaches 2^32, the situation becomes worse.  The problem is that a reduce function that takes only the lower 32 bits of the hash cannot evenly distribute to more than 2^32 buckets.

Since the updated hash function already relies on Intel/AMD/ARM u64xu64->u128 instruction, there seems to be no harm in using that instruction in reduce, allowing it to take u64 hashes instead of u32.

The reduce function is used rather than the standard masking operation on the hash to derive a table index because we did not want to waste the space required to round the table size up to the next power of 2.  Reduce works on any table size.