Improve distribution of reduce function. (#4854)
* Enhanced hash_u64 as 2-rounds.

* Slight improvement to the hash fucntion.

By XORing the input value into the outpout only at the end, we maintain a permutation until that step.  This ensures no information is lost in the intermediate steps.  With this scheme, a 3-round version passes the diehardet tests when ouitputing all bits of the hash when the input is a counter.

* Switched to folding multiply, similar to aHash, but with some more mixing before.

This now passes distribution tests, cycle length tests, and is just as fast as aHash's fallback path for u64.  The current aHash code still fails distribution tests, but if the author adopts something like what is in this PR, we could switch to aHash.

* Changed hash_secret to a more random value to pass tests.

* Fix for debug mode overflow

* Change the reduce function to use 128-bit multiply.

The reduce function before this change did not uniformly select a hash table location, slightly increasing the number of collisions when inserting a billion entries into the table.  As the number of keys approaches 2^32, the situation becomes worse.  The problem is that a reduce function that takes only the lower 32 bits of the hash cannot evenly distribute to more than 2^32 buckets.

Since the updated hash function already relies on Intel/AMD/ARM u64xu64->u128 instruction, there seems to be no harm in using that instruction in reduce, allowing it to take u64 hashes instead of u32.

The reduce function is used rather than the standard masking operation on the hash to derive a table index because we did not want to waste the space required to round the table size up to the next power of 2.  Reduce works on any table size.

* Reformatted lookup_htbl.rs.

* Updated variable names in reduce to be more clear.

* Changed table_len to usize.