Initial port of Chromium's noise_nk handshake.
This is a quick and dirty port of https://source.chromium.org/chromium/chromium/src/+/main:third_party/cloud_authenticator/?q=third_party%2Fcloud_authenticator&ss=chromium.  Multiple upgrades are needed in follow-on PRs:

* Add client-side code, basically just copying from tests.
* Clean up rustcrypto.rs interface to the crypto used.
* Add support for noise_kk.

This is server-side code, but the client side code can be gleaned from tests.  The rustcrypto.py file is derived from code influenced by the Ring crypto API, and ideally we should simplify and clean up the interface.  It was meant to be used for linking to different crypto backends, but as Ring was first, rustcrypto.rs is basically a hack to conform to Ring's API.  This PR adds initial noise-nk support, but for enclaves talking to other enclaves, we probably need noiew-kk, which should be added.