Revert Linux kernel to version 6.1.33 (#4915)
Version 6.7.6 is not compatible with stage 0 on SEV, SEV-ES and SEV-SNP. Version 6.8 is also not compatible. It looks like 6.9 should be compatible once it is released, so we can upgrade then. Reverting for now to an older version we know was compatible untill we can upgrade to a newer compatible version.

See b/327367706