Improve oak containers system base image reproducibility.
Previous changes focused on making the contents of `base-image.tar`
reproducible; this change focuses on making the tarball itself
reproducible, assuming the same contents. This means that the entries in
`base-image.tar` need to have deterministic (a) a order, (b) timestamps,
and (c) permissions.

As part of this change, we switch to using BuildKit instead of the
legacy Docker builder because it supports `COPY --chmod`.

There are no changes to the files within `base-image.tar`. The tarball
itself has several changes:

* Timestamps are different
* No longer two copies of `etc/hosts`, and `etc/hosts` is now owned by
  root, not the current user.
* Several config files in `/etc` now have mode 755 instead of 644; this
  is unfortunately necessary when using `COPY --chmod` to avoid removing
  execute permissions from directories. These files include
  * `etc/systemd/journald.conf.d/override.conf`
  * `etc/systemd/network/10-enp0s1.network`
  * `etc/systemd/system/oak-orchestrator.service`
  * `etc/systemd/system/oak-syslogd.service`
  * `etc/systemd/system/root-passwd.service`
  * `etc/systemd/system/serial-getty@.service.d/override.conf`
  * `etc/sysusers.d/oak.conf`

The permission changes are unfortunate but consistent with a number of
other config files in `/etc` -- probably for the same reasons. But
unlike the previous `chmod a+rX` approach, the resulting permissions do
not inherit the write permissions from the build environment, which
likely depend on the user's umask settings.

Bug: 332535604
Change-Id: Ibe44195afbb8395b0bf1fecfd1c2b425ca7295dd