Remove some seemingly unnecessary `Arc`-s
I was looking at this code for unrelated reasons and one thing that got
me thinking was our overuse of Arc-s everywhere. This replaces some uses
with `Box`es and plain borrows.

This should be a no-op, aside for saving a few instructions as we don't
have to maintain an atomic counter every time we call `handle_invoke`.
But that was not the reason for this change (it's just a cleanup), and
I'm pretty sure we won't have any noticeable performance improvements.

Change-Id: I25d242cf7dec264af00cd9fc8e565861559687dd