Add a config struct for the wasm engines
This is not in use right now, but will allow us to dynamically change
properties about the wasm engine in the future.

The main motivation is to provide a way how to tweak the wasmtime memory
parameters dynamically without having to recompile the whole binary, as
some might depend on how large of a wasm module we intend to load.

Change-Id: I8e02d9c79abc65966ee4299c683a2180dd5400e1