Create a `WasmtimeConfig` proto that exposes wasmtime memory settings
This will let us dynamically tweak them, if needed. The best values will
depend on the exact workload, hence it's better to have them dynamically
adjustable instead of just baked into the binary. (If you don't know
what good values would be, you can always leave them with default
values.)

As there's _a lot_ of fields, I also introduced a tiny macro to make it
less tedious.

I'm hoping that this could help in the future with the wasmtime memory
allocation overhead, but that's going to be a separate undertaking as
there will need to be a lot of testing and benchmarking of these
settings. (And in the end we may come to the conclusion that it didn't
help.)

Bug: 334075655
Change-Id: Iea66822374d7da1dae75d295dbc3753c263dfb53