Add tests for explicit reference value verification
These tests extract the actual values from the evidence as reference values to make sure that the attestation verification works with explicit reference values for all fields. The one exception is the application layer configuration for the restricted kernel. It is not currently specified, so must be skipped for now.

Adding tests that modify the correct reference values to test failure will be added as a follow change.

BUG: 325509208

Change-Id: Idffe34f9bb4adb715efefeaf6b947209dd5c2568