Use Bazel remote cache
This lets everyone read from the cache, but in order to write to it it
is necessary to have a valid service account and key locally, which
currently needs to be done manually.

This will at least allow us to see if remote builds are faster given the
current cache, and then decide on next steps.

Move CI related bazel flags to the local .bazelrc, and use the new `ci`
config when running in CI.

Remove unused cloudbuild file.

Bug: 337266665
Change-Id: I9da558b4338b9558f83e8a77e3798d0c0beffe06