Placeholder implementation of reference value descriptions.
Adds a simple implementation for now (literally just the display string), so other work can proceed. Will be improved later.

Long term I'm a little bit skeptical creating manual descriptions for each reference values, as they risk becoming comments describing the functioning of the verification library that may become inaccurate. Eg see http://b/337920594 as an example of how behavior is unclear from reference values alone.

The better approach may be to implement a slightly prettier version of the std::fmt::Display for the relevant structs (eg one that omits fields set to None) and rely on the technical expertise of the auditor to make sense of these printed values.

Bug: 334900893

Change-Id: I67e9ed63b9e3501c8baa52ad743adf3655389e1a