Add support for transferring lookup data in a serialized format
The hope here is that this will reduce allocator pressure. One chunk may
contain ~60K entries, so the allocator has to do a fair amount of work
for the prost deserializer (allocate a vector for the entries, allocate
memory for all the fiels inside entries, fill them, ...) only for us to
copy data to the hash map and then release the memory fairly quickly.

This will let us deserialize one message at a time, hopefully relying
more on the heap. The downside is that we'll likely hold on to the mutex
longer; but we already have discussed about splitting that mutex up as
well.

The new implementation is also completely backwards compatible; the
`oneof` doesn't change how data is serialized.

Preliminary tests suggest that it is indeed consistently about ~20%
faster, but that's on my workstation with a limited data set size. I'll
need to do some more large-scale testing, but the first results are
promising.

Bug: 337854886
Change-Id: Ie3a1e2a996ec9cbc851e144db88aacf58e376de9