Put generated Prost code in an importable location
During the oak_proto_rust build script, the generated proto files also get
copied to a known location inside of the source tree.

We can automatically import these files using copybara.

A CI check is added to make sure that the generated files are included in the
expected location.

This removes the need to manually generate/copy files when proto changes are
made.

Change-Id: I7351d5348b3fec22824faf26a8d99eabdbbf52ae