Add Encryptor trait
This trait is a part of Low-Level Oak SDK and will be used by Trusted Brella to implement their custom encryptor that should work with Oak protocol.

We are also planning to use this trait to implement both Noise and Hybrid encryptors (Hybrid will be refactored from an existing implementation to fit the new Low-Level SDK).

Bug: 338559159
Change-Id: I57d9405ccf7a9024c6a2f3624f6b9d2520619698