Ensure the todo linter works okay with across mulitple lines.
By no longer matching that the line must end in a period.

This does allow TODOs that do not end in a period, even if the TODO is not broken up among multiple lines.

Imo this the desired behavior. The current check is flaky and not worth it. It already only applies to one form of the accepted todos (a bug). We could try to make it work across multiple lines but it'd get into the territory of overengineering quickly, as that leads to lots of subtle problems such as dealing with indentation, differentiating comments and code.

Bug-Id: 333067027
Change-Id: I52d5c9e974d5944c6825e557d90f1f517f4be400