Add support for connecting to launcher via vsock to stage1
The launcher will listen on both vsock and TCP, as other things (like
the orchestrator) don't know how to use vsock.

This is opt-in via the command line flag; by default, stage1 still tries
to use the network. The main goal is to test virtio-vsock with the
connection initiated by the guest, not the host.

(We should get comparable performance numbers using the base image with
nvidia drivers; it's 4 GB so it should take a bit to load.)

This may let us simplify the initial boot procedure as well, as now we
don't have to bring up the network interface during boot for stage1, but
rather can delegate any and all network business to systemd-network once
we've loaded the system image.

Bug: 289334144
Change-Id: If714c69bb4e69d0d9da8eaccf44fd424fb855187