Use a more consistent way of finding well-known protos for build scripts
The old approach was based on me just digging around in the sandbox
directories and finding files.

Hopefully this approach is better: it uses the proper runfiles resolution
helper library to find included data files.

This also takes advantage of the well-known types target in the protobuf
library, handling most of what we need (but we still need to find
descriptor.proto individually).

Change-Id: I3c7f2352f8adc39b36e5a72063956af5911b7132