Rely on baze make variable `location` to find protobuf include paths.
Similar change was applied to oak_proto_rust.

The current approach assumes that protobuf is included under com_google_protobuf.
However, that might not be always the case as it might be protobuf~ if included from MODULE.bazel
or protobuf if included from WORKSPACE.

This change would allow repositories depending on `oak` not rename the imports.

Bug: b/344012112
Change-Id: I4bd43ca5dc236645e93e927c4771d3222cae3522