Oak Standalone: Milestone 1: No-Op attestation PoC
In milestone 1 of Oak Standalone, we'll focus only on estabilshing the
crypto channel to communicate with a demo server, see: go/oak-standalone

As a first step, this CL introduces the no-op attestation flow that
provides the encryption PK without requiring any attestation
verification.

The primary goal of the CL is to lay out some boilerplate and set us up
for quick iterative improvements on the standalone concept.

The facade is based on the existing stream_session.proto service API.

Key elements:
* Create the oak_standalone crate
* Create a standalone client crate that instantiates a new OakClient
  using a special standalone AttestationVerifier.
* Create a standalone service that can establish an encrypted channel
  with an OakClient using the stream sesion protocol.
* Add a "Standalone" mode to endorsements. It's mostly ignored for now
  but may be more useful later.

Fixed: b/347967540

Change-Id: I8f789fcbb061dcafc3c7bf696367f578aae6e913