Store userspace context in proper struct when handling syscalls
The idea here is that once we have the userspace context stored in a proper struct (instead of on the temporary kernel stack), we can later persist it and resume paused processes.

It's the first time I'm writing assembly (http://memegen/9887503231543599), so for now this just leaves the existing logic for saving registers as is. We should replace it with xsave later on probably.

Change-Id: Id9887487e4fd2ec36206359cd0050bab2c989df3