Build (and install) nvidia drivers in the nvidia_base_image
This is a bit hacky as we effectively rebuild the whole kernel while creating
the system image so that we could get the kernel-headers package, which is
required for DKMS to build the nvidia drivers.

We don't use the built kernel but rather rely on the one nix built; this does
trigger some warnings as the nix-built kernel and the nvidia modules have been
built using different compilers which is dangerous territory; at some point in
the future we need to see how we could build both the kernel and the modules in
the same environment. (Which likely means not using nix to build the kernel.)

For now, the modules can be loaded just fine and `nvidia-smi` detects the GPU.

Bug: 289334314
Change-Id: I4b90b7207429e7729af74cd87c79ea8986f1de11