Add a `verify` variant that returns expected values
Introduces `verify_and_return_expected_values`, which is just like the
existing `verify` call, but with a different return type that also
includes the proto containing all computed expected values.

For consistency, the existing `verify` function maintains the existing
shape. Eventually we may want to refactor things so that only the
`verify` call remains, and it has the new API shape.

Now that we have an API that returns the expected_values, we can add tests to
verify that they are correctly populated, so one simple test to confirm that
the expected expected values are present is included here.

Bug: 333581670

Change-Id: Id2c365de886e4985ece7b990da8d683959b66959