Replace regex crate in oak_attestation_verification by regex-lite 0.1.6.
This will make it possible to enable the `regex` feature in oak_attestation_verification also in cases where the the size of the binary is critical (see bug for measurements).

The downside is that even though regex-lite has a `std` feature, it cannot be disabled so far. Fixing this is on the list of the maintainer, and is also feasible. Apparently it is mostly a dependency on std::sync::Mutex. If it becomes blocking for us we could contribute to regex-lite.

Bug: 336723429

Change-Id: Ie69a552a6ebf3cf5cb02a062e60bb3033c0e4e63