Bazelify micro_rpc_tests
Fixes: b/349587495

Since nothing depends on it, we can also de-cargo it!

Uses the `write_source_files` bazel rule to manage copying the generated
file to the example location.

The :build target only exposes the entire out directory to dependencies,
so the only option I could find was to copy the entire directory. I
copied it to a location named `generated` so it gets ignored by linter
tools, etc. This removes the need to rename the file to .txt

Change-Id: I8eeeb9bccfba5eb7637e0cb8db0e98c6df3f08a3