Purge (sometimes forcefully) some packages from the nvidia system image
There's stuff in there we know we won't need, like the compilers, as we
know we'll never try to recompile the kernel modules.

This reduces the size of the compressed system image by ~150M (390->243).

Unfortunately it's rather unlikely we'll be able to reduce it further
that much, as some of the nvidia libraries themselves are a good 80M
apiece, and we're going to need them.

I reckon we could get it down to 200ish M compressed, but that'd be
really scraping the barrel.

Bug: 298933839
Change-Id: Ic5237f5b1cd4005a1a8748439c6417cc28aae17a