Add TDX metadata GUID block and sections to stage0_bin
This is the first change for TDX enlightenment in stage0. It adds
a GUID block for TDX Metadata Offset, and a 32-byte section header
for TDX Metadata.

Before this change, the hypervisor fails because of missing TDX
metadata offset block. After this change, the hypervisor fails at
TD_HOB validation, which means the metadata offset block and the
section headers are both in good shape.

Bug: 352784173
Change-Id: Id809a4ef5933a9958b619267dfbce6966c7eedf0