Build all gRPC interfaces in one place
It's been tricky to add gRPC services to oak_proto_rust, but we can at
least create a separate, single build script/crate that builds all gRPC
services.

This removes lots of repeated generation of the same Prost code in lots
of different packages. It removes a large number of build scripts, so
that we don't have to keep track of as many.

Changes in this CR:
* Create oak_grpc, with a single build script that builds all services.
* Consistently use oak_proto_rust version of protos everywhere.
* Rust package oak_functions -> functions to match proto package
* Remove all per-crate `mod proto` hierarchies, use either
  `oak_proto_rust` or `oak_grpc`

Change-Id: I95647b3f559c97bfc3ae76c27a8165f02018d593