Add justfile recipe to run tests
This new recipe ensures that all the dependencies are rebuilt and up to
date.

As an example:

Modify
/usr/local/google/home/tzn/src/oak/oak_restricted_kernel_wrapper/src/asm/boot.s
and add an `int3` instruction in `_wrapper_entry`, then run `just
run_oak_functions_test`, and see it fail. Then remove the `int3` and
rerun the command, and see it pass again within 30s.

Bug: 355606262
Change-Id: Idbc4c42c3f926129d55103349b49dfe9f25b6f3c