Add a demo HTTP Transport option to the hello_world example
This is a small step towards a web-based Oak containers demo. To make it
easier to interact with an Oak server from a web browser, we add an
HTTP-based transport to the hello world example.

This transport is not meant for general use yet, it's just a quick
implementation for demo purposes.

Fixed: b/355220506

Change-Id: Id68c9ff66f7892233f95e391f87bfc4e56b5c953