Fully bazelify hello world trusted app
We remove the need for a separate container directory that does the
docker build: we can build with trusted rust app with just a few simple
OCI rules now.

Since nothing depends on the trusted app, the Cargo.toml is removed, and
it can only be build with bazel now.

Change-Id: I519b4069ec64d5c9f9ae0baa68cd2afab774dee0