Provide endorsed evidence from Orchestrator
Add ability to the orchestrator to handle GetEndorsedEvidence requests,
so that the client's TEE app can handle these requests as part of the
same stream.

* The orchestrator now holds on to a copy of the evidence it creates,
  but still sends it to the launcher for compatibility.
* Uses the newly added launcher endpoint to get endorsements.
* The orchestrator ipc service exposes a get_endorsed_evidence
  call. The TEE app can use this on startup so that it can handle
  get_endorsed_evidnece calls.

This simplifies the hostlib server implementation, as it becomes a
simple proxy, with all complex request/session logic residing in the
trusted app.

A lot of the logic for Oak-related things can eventually be moved into
SDK functions that these examples will use.

Fixed: b/338012560

Change-Id: I485418272aa89a677e1f29b913ae229e579158fe