Refactor stage0 GHCB operations into a struct
Short term, I think this makes the code a bit more Rust-y.

Longer term, my overarching idea is (a) generalize `SharedAllocator` and
move it out of `sev.rs` as we'll likely need a similar concept for TDX,
(b) make the GHCB code use the `SharedAllocator` instead of managing the
page state manually, and (c) move the page state operations
(`share_page`, `unshare_page`) to the HAL and thus it helps if it's only
the `SharedAllocator` is the one that manages the sharedness state.

Change-Id: I2e8cefaa67f7f111026811466624febfd31bcc62