Modify opentelemetry_sdk to be no_std compatible (part 3)
Turns out that until now, lots of targets were getting skipped when running tests. Some would fail when actually run, so this fixes the failures. This tweaks the tests so they're not reliant on the async runtime. Unfortuantely, the inmemoryexporter and manual reader don't exactly work together for collecting/reading metrics, so we need to use a reader directly for testing

- Edit tests to use reader directly over exporter

Change-Id: Ibc1a618deee7de9b4e15945aa4996990698d2a00