Don't drag the encrypted bit around in `paging.rs`
The `encryped()` facility will need to be replaced with something more
generic soon anyway as that's not how TDX works, but that'll be in a
separate CR.

Also, while I'm in there, store `pt_0` in a `Box` instead of just
leaking it. Functionally it's the same as we put the `Box` in a static,
but it's a bit cleaner as it still nominally tracks ownership properly.

Change-Id: I23ab3b94b5a3f599bb54467b05c9a4e61ae5f71f