Make the page encryption intent explicit.
Previously we just OR-ed the encryption bit to the physical address
whenever needed, but this won't work with TDX as the bits work the other
way around. Therefore, let's add a new interface that lets you deal with
page table entries and the encryptedness of said page explicitly.

Next steps are to generalize the handling of the encrypted bit out of
`paging.rs` and into the HAL.

One downside with the current implementation is that you can still call
the underlying `set_addr()`/`addr()` directly, netting you an PhysAddr
with the encryptedness state bit set unexpectedly. However, that'll
likely mean reimplementing `PageTable`/`PageTableEntry` ourselves, and
I'm questioning whether that's worth the trouble.

Bug: 350496083
Change-Id: Id7d19773749a35429459523cd156d6ca44568ea6