Introduce wrapper for PageTableEntry that doesn't expose `set_addr()`
Optional follow-up to the previous CR; feel free to push back on the
change.

We don't want people to use `addr()`/`set_addr()` directly as those
functions don't know anything about the encrypted bit; however, you
can't just hide a method of a struct without resorting to tricks like
these.

The ugliness is the unsafe blocks that cast raw pointers around; it's
effectively doing a transmute with all the safety risks it entails.

Bug: 350496083
Change-Id: I9d877b6c5f6e93efed0221ad7ae11b62b1f3c9af