Clean up GHCB deinitialization code (and mark it unsafe as it is).
There are some contentious changes here, especially around adding the
unsafe `deinit()` to OnceCell, so feel free to push back there.

My longer-term idea is that instead of `Box::leak` and managing the
sharedness of the page manually, we'd use the existing `Shared` struct
to manage the page state: in `Ghcb::init()` we'd do `Shared::leak`
(mirroring Box::leak) and in `deinit()` we'd reconstruct the `Shared`
with `Shared::from_raw_in`, letting `Shared::drop` mark the page as
unshared as we exit `deinit`.

This would consolidate the page sharedness management solely into the
`Shared` struct.

Ideally we'd be able to feed `Shared<Ghcb>` straight into `GhcbProtocol`
and get it back with `into_inner()` later, no `leak()` nor `from_raw_in`
required, but let's just say the type signature we've ended up with for
`GhcbProtocol` is rather arcane and hard to generalize like that.

Change-Id: Ib5c07fb9e1ee8c65cbde1ed292c20b306c2590ef