Build stage0_bin with Bazel
Apologies for the large CR. I couldn't figure out a good way to break
it into smaller chunks while keeping it testable.

Stage 0 requires a large code model when building, which in turn
requires rebuilding the std libraries. All the changes in the root
WORKSPACE file and the `bazel` directory are related to rebuilding
the std libraries. This includes downloading the sources, adding a
build file, building the libraries and then setting up a new toolchain
to use for building Stage 0.

The code for rebuilding the std lib is based on the example in
https://github.com/daivinhtran/rules_rust/blob/toolchain-to-rebuild-std/examples/toolchain-to-rebuild-std/

Bazel requires that the linker arguments are passed differently in the
build script, and the linker we are using also required a reordering of
the linker script.

The linker script also needed multiple fixes to be compatible with both
LD and LLD.

Unfortunately I had to also create a new crate index to support building
the `sha2` crate with the `force-soft` feature enabled. Crate indices
don't support `select` statements, so this could not be configuration
based. I tried to clean it up a bit so we don't have to maintain
two separate no_std crate index definitions.

BUG: 349587224

Change-Id: I97f46891da427829586394895b977f15537b0b12