Use the `Shared` struct to manage the GHCB.
This delegates the page sharedness management to `Shared`, so that we
don't have to manage that ourselves.

Unfortunately we do have to first leak the memory and then later
reassemble it ourselves; ideally we'd be able to pass ownership of the
whole `Shared` to `GhcbProtocol` (so that we could get it back at deinit
time), but that's not possible right now and will be a significant
refactoring.

Change-Id: I0f8170461ac8d84c5cb27c6652577bfdbceba5d5