Remove the hard coded page tables
Currently the guest VM starts with a hardcoded page table to jump
to 64-bit long mode, and set up another one in the low memory area,
and we accept a 2MiB page for the low memory area. However, Linux
kernel may not like it if the memory between 640KiB and 1MiB was
accepted.

In this CR we use the 640KiB TEMP_MEM section in the low memory
area for the page tables. The hypervisor adds it and guest can use
it without doing tdcall. So we can dynamically set up the page
tables in the .bss section.

Change-Id: I7dfc18304e91fbbb7849d2bd2fbc1784ba5bae91