Move the `SharedAllocator` to `allocator.rs` and delete `sev.rs`.
Everything has moved our of `sev.rs` to either the HAL or some other
generic place, so there's nothing left in there.

This should conclude the main points of HALification of stage0. There is
some clean-up work still to be done as there are unconditional
dependencies on some SEV-related crates, especially for data structures
that should be genericised, but that's not going to block the TDX work.

Bug: 350496083
Change-Id: I9a0204e8effe18522f4b29b8003a56f13110b5de