Start creating a trait for the HAL
The main difficulty stemmed from the fact that we don't want the
platform to leak into the `PortFactory` or `Port` types, as we need to
have the serial port as a static and at that point we don't know what
the right `Platform` is yet. The answer is... function pointers, as they
have a known size and signature.

This just covers CPUID + MMIO + ports, but this should suffice as an
example. I will move the rest of the operations (MSRs, attestation etc)
to the trait in a separate CR as those changes are more
straightforward.

Bug: 350496083
Change-Id: Ide1e100147a9463c44b664edc9856d4e78151e00